remove redundant view models
