+++ app/src/main/java/ngo/teog/swift/gui/main/BarcodeFragment.java
@@ -49,7 +49,7 @@ public class BarcodeFragment extends Fragment {
     @Inject
     ViewModelFactory viewModelFactory;
 
-    private BarcodeViewModel viewModel;
+    private MainViewModel viewModel;
 
     private DecoratedBarcodeView barcodeScannerView;
     private String lastText;
@@ -105,8 +105,8 @@ public class BarcodeFragment extends Fragment {
 
         progressBar = view.findViewById(R.id.progressBar);
 
-        if(ContextCompat.checkSelfPermission(this.getContext(), Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
-            ActivityCompat.requestPermissions(this.getActivity(), new String[]{Manifest.permission.CAMERA}, 0);
+        if(ContextCompat.checkSelfPermission(this.requireActivity(), Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
+            ActivityCompat.requestPermissions(this.requireActivity(), new String[]{Manifest.permission.CAMERA}, 0);
         }
 
         DaggerAppComponent.builder()
@@ -115,7 +115,11 @@ public class BarcodeFragment extends Fragment {
                 .build()
                 .inject(this);
 
-        viewModel = new ViewModelProvider(this, viewModelFactory).get(BarcodeViewModel.class);
+        SharedPreferences preferences = this.requireActivity().getSharedPreferences(Defaults.PREF_FILE_KEY, Context.MODE_PRIVATE);
+        int id = preferences.getInt(Defaults.ID_PREFERENCE, -1);
+
+        viewModel = new ViewModelProvider(requireActivity(), viewModelFactory).get(MainViewModel.class);
+        viewModel.init(id);
     }
 
     @Override
@@ -136,11 +140,8 @@ public class BarcodeFragment extends Fragment {
         searchButton.setVisibility(View.INVISIBLE);
         progressBar.setVisibility(View.VISIBLE);
 
-        SharedPreferences preferences = this.getContext().getSharedPreferences(Defaults.PREF_FILE_KEY, Context.MODE_PRIVATE);
-        int userId = preferences.getInt(Defaults.ID_PREFERENCE, -1);
 
-        viewModel.init(userId, deviceId);
-        viewModel.getDeviceInfo().observe(BarcodeFragment.this.getViewLifecycleOwner(), deviceInfo -> {
+        viewModel.getDeviceInfo(deviceId).observe(BarcodeFragment.this.getViewLifecycleOwner(), deviceInfo -> {
             if(deviceInfo != null) {
                 Intent intent = new Intent(BarcodeFragment.this.getContext(), DeviceInfoActivity.class);
                 intent.putExtra(ResourceKeys.DEVICE_ID, deviceInfo.getDevice().getId());
