+++ app/src/main/java/ngo/teog/swift/gui/main/MainViewModel.java
@@ -3,13 +3,18 @@ package ngo.teog.swift.gui.main;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.ViewModel;
 
+import java.util.List;
+
 import javax.inject.Inject;
 
+import ngo.teog.swift.helpers.data.DeviceInfo;
 import ngo.teog.swift.helpers.data.Hospital;
 import ngo.teog.swift.helpers.data.HospitalRepository;
 
 public class MainViewModel extends ViewModel {
+    private int userId;
     private LiveData<Hospital> hospital;
+    private LiveData<List<DeviceInfo>> deviceInfos;
     private final HospitalRepository hospitalRepo;
 
     @Inject
@@ -22,10 +27,24 @@ public class MainViewModel extends ViewModel {
             return;
         }
 
+        this.userId = userId;
         hospital = hospitalRepo.getUserHospital(userId);
+        deviceInfos = hospitalRepo.getHospitalDevices(userId);
     }
 
     public LiveData<Hospital> getUserHospital() {
         return hospital;
+    }
+
+    public LiveData<DeviceInfo> getDeviceInfo(int deviceId) {
+        return hospitalRepo.getDevice(userId, deviceId);
+    }
+
+    public LiveData<List<DeviceInfo>> getDeviceInfos() {
+        return deviceInfos;
+    }
+
+    public void refreshHospital(int userId) {
+        hospitalRepo.refreshUserHospital(userId);
     }
 }
